GM6-Aircraft
Añade la entidad correspondiente al requisito Mandatory 6 (Aircraft).
