+++ src/main/java/acme/entities/airline/Airline.java
@@ -1,10 +1,15 @@
 
-package acme.entities.group;
+package acme.entities.airline;
+
+import java.util.Date;
 
 import javax.persistence.Column;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Past;
 import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Size;
 
@@ -13,7 +18,7 @@ import org.hibernate.validator.constraints.URL;
 
 import acme.client.components.basis.AbstractEntity;
 
-public class Airport extends AbstractEntity {
+public class Airline extends AbstractEntity {
 
 	// Serialisation identifier
 	private static final long	serialVersionUID	= 1L;
@@ -25,30 +30,27 @@ public class Airport extends AbstractEntity {
 
 	@Column(unique = true)
 	@NotBlank
-	@Pattern(regexp = "^[A-Z]{3}$", message = "{validation.airport.code}")
+	@Pattern(regexp = "^[A-Z]{2}X$", message = "{validation.airline.code}")
 	private String				code;
-
-	@NotNull
-	private OperationalScope	operationalScope;
-
-	@NotBlank
-	@Size(max = 50)
-	private String				city;
-
-	@NotBlank
-	@Size(max = 50)
-	private String				country;
 
 	@URL
 	@Length(max = 255)
 	private String				website;
 
+	@NotNull
+	private AirlineType			type;
+
+	@NotNull
+	@Past
+	@Temporal(TemporalType.DATE)
+	private Date				foundationMoment;
+
 	@Email
 	@Length(max = 100)
 	private String				email;
 
-	@Pattern(regexp = "^\\+?\\d{6,15}$", message = "{validation.airport.contactPhoneNumber}")
-	private String				contactPhoneNumber;
+	@Pattern(regexp = "^\\+?\\d{6,15}$", message = "{validation.airline.phoneNumber}")
+	private String				phoneNumber;
 
 	// Derived attributes
