+++ src/main/java/acme/entities/airline/AirlineType.java
@@ -1,5 +1,5 @@
 
-package acme.entities.group;
+package acme.entities.airline;
 
 public enum AirlineType {
